+++ src/main/java/com/revature/krowdboot/utility/EmailValidator.java
@@ -1,0 +1,17 @@
+
+package com.revature.krowdboot.utility;
+
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+
+public class EmailValidator  {
+  private static final String EMAIL_PATTERN = "^[_A-Za-z0-9-+]+(.[_A-Za-z0-9-]+)*@[A-Za-z0-9-]+(.[A-Za-z0-9]+)*(.[A-Za-z]{2,})$"; 
+  private static final Pattern pattern = Pattern.compile(EMAIL_PATTERN);
+  
+  public static boolean isValid(String email) {
+      Matcher matcher = pattern.matcher(email);
+      return matcher.matches();
+  }
+  
+}
